+++ internal/origin_leaf/leaf/tokens.go
@@ -15,12 +15,14 @@ func (s *Session) DeleteToken(id string) {
 }
 
 // delete the space on all leaf nodes
-func DeleteToken(id string) {
+func DeleteToken(id string, skipSession *Session) {
 	sessionMutex.RLock()
 	defer sessionMutex.RUnlock()
 
 	// Send the user to all followers
 	for _, session := range session {
-		session.DeleteToken(id)
+		if session != skipSession {
+			session.DeleteToken(id)
+		}
 	}
 }
